+++ LICENSE.txt
@@ -1,6 +1,6 @@
 The MIT License
 
-Copyright (c) 2021-2022 Check Digit, LLC
+Copyright (c) 2021-2023 Check Digit, LLC
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
