+++ package.json
@@ -27,10 +27,10 @@
     "prettier": "prettier --ignore-path .gitignore --list-different .",
     "prettier:fix": "prettier --ignore-path .gitignore --write .",
     "test": "npm run ci:compile && npm run ci:test && npm run ci:style",
-    "build-tsc": "rimraf build && tsc && node --es-module-specifier-resolution=node ./build/index.js 2>&1 | grep -q 'complete'",
-    "build-es": "rimraf build-es && esbuild ./src/* --platform=node --bundle --format=esm --sourcemap=inline --outdir=build-es && node ./build-es/index.js 2>&1 | grep -q 'complete'",
-    "build-swc": "rimraf build-swc && swc ./src -d ./build-swc && node --es-module-specifier-resolution=node ./build-swc/index.js 2>&1 | grep -q 'complete'",
-    "ci:test": "rimraf build build-es build-swc && NODE_OPTIONS=\"--no-warnings --experimental-vm-modules\" jest --coverage=false",
+    "build-tsc": "rimraf build && tsc",
+    "build-es": "rimraf build-es && esbuild ./src/* --platform=node --bundle --format=esm --sourcemap=inline --outdir=build-es",
+    "build-swc": "rimraf build-swc && swc ./src -d ./build-swc",
+    "ci:test": "NODE_OPTIONS=\"--no-warnings --experimental-vm-modules\" jest --coverage=false",
     "ci:compile": "npm run build-tsc && npm run build-es && npm run build-swc",
     "ci:style": "npm run prettier"
   },
@@ -40,12 +40,35 @@
     "@swc/core": "^1.3.32",
     "@types/jest": "^29.4.0",
     "esbuild": "^0.17.5",
-    "esbuild-jest": "^0.5.0",
     "get-port": "^6.1.2",
-    "got": "^12.5.3",
+    "got": "^11.8.6",
     "jest": "^29.4.1",
     "rimraf": "^4.1.2",
     "ts-jest": "^29.0.5"
+  },
+  "jest": {
+    "extensionsToTreatAsEsm": [
+      ".ts",
+      ".mts"
+    ],
+    "moduleNameMapper": {
+      "^(\\.{1,2}/.*)\\.js$": "$1"
+    },
+    "transform": {
+      "^.+\\.m?[tj]sx?$": [
+        "ts-jest",
+        {
+          "useESM": true
+        }
+      ]
+    },
+    "collectCoverageFrom": [
+      "<rootDir>/src/**",
+      "!<rootDir>/src/**/*.spec.ts"
+    ],
+    "testMatch": [
+      "<rootDir>/src/**/*.spec.ts"
+    ]
   },
   "overrides": {
     "ts-jest": {
