+++ src/module-directory/index.ts
@@ -1,3 +1,3 @@
 // module-directory/index.ts
 
-export default () => 'module-directory';
+export default () => 'module-directory-index';
