+++ src/module.ts
@@ -1,0 +1,9 @@
+// module.ts
+
+export function test() {
+  return 'export function test';
+}
+
+export default function () {
+  return 'export default function';
+}
