+++ src/typescript-4.6.spec.ts
@@ -1,0 +1,20 @@
+// typescript-4.6.ts
+
+import { strict as assert } from 'node:assert';
+
+describe('typescript-4.6', () => {
+  it('has typescript 4.6 features', () => {
+    // control flow analysis for destructured discriminated unions
+    type Action46 = { kind: 'NumberContents'; payload: number } | { kind: 'StringContents'; payload: string };
+
+    function processAction46(action: Action46) {
+      const { kind, payload } = action;
+      if (kind === 'NumberContents') {
+        assert.ok(payload * 2);
+      } else if (kind === 'StringContents') {
+        assert.ok(payload.trim());
+      }
+    }
+    assert.equal(processAction46({ kind: 'NumberContents', payload: 5 }), undefined);
+  });
+});
