+++ src/typescript-4.7.spec.ts
@@ -1,0 +1,17 @@
+// typescript-4.8.ts
+
+import { strict as assert } from 'node:assert';
+
+describe('typescript-4.7', () => {
+  it('has typescript 4.7 features', () => {
+    // control-flow analysis for bracketed element access
+    const key = Symbol();
+    const numberOrString = Math.random() < 0.5 ? 42 : 'hello';
+    const obj = {
+      [key]: numberOrString,
+    };
+    if (typeof obj[key] === 'string') {
+      assert.ok(obj[key].toUpperCase()); // 4.7 knows that obj[key] is a string
+    }
+  });
+});
